Move TrustStoreManager from GUI to Logic
